Moved intrinsics to its own module
