+++ src/lifetime.rs
@@ -7,6 +7,8 @@ use std::cmp::{PartialOrd, Ordering};
 use self::piston_meta::MetaData;
 use self::range::Range;
 
+use intrinsics::{self, ArgConstraint};
+
 pub fn check(data: &[Range<MetaData>]) -> Result<(), String> {
     let mut nodes: Vec<Node> = vec![];
     let mut parents: Vec<usize> = vec![];
@@ -220,38 +222,7 @@ pub fn check(data: &[Range<MetaData>]) -> Result<(), String> {
     }
 
     // List all intrinsic functions.
-    // Stores name, number of arguments, returns
-    let mut intrinsics: HashMap<&'static str, Intrinsic> = HashMap::new();
-    intrinsics.insert("println", PRINTLN);
-    intrinsics.insert("print", PRINT);
-    intrinsics.insert("clone", CLONE);
-    intrinsics.insert("debug", DEBUG);
-    intrinsics.insert("backtrace", BACKTRACE);
-    intrinsics.insert("sleep", SLEEP);
-    intrinsics.insert("round", ROUND);
-    intrinsics.insert("random", RANDOM);
-    intrinsics.insert("read_number", READ_NUMBER);
-    intrinsics.insert("read_line", READ_LINE);
-    intrinsics.insert("len", LEN);
-    intrinsics.insert("push", PUSH);
-    intrinsics.insert("trim_right", TRIM_RIGHT);
-    intrinsics.insert("to_string", TO_STRING);
-    intrinsics.insert("typeof", TYPEOF);
-    intrinsics.insert("sqrt", SQRT);
-    intrinsics.insert("sin", SIN);
-    intrinsics.insert("asin", ASIN);
-    intrinsics.insert("cos", COS);
-    intrinsics.insert("acos", ACOS);
-    intrinsics.insert("tan", TAN);
-    intrinsics.insert("atan", ATAN);
-    intrinsics.insert("exp", EXP);
-    intrinsics.insert("ln", LN);
-    intrinsics.insert("log2", LOG2);
-    intrinsics.insert("log10", LOG10);
-    intrinsics.insert("random", RANDOM);
-    intrinsics.insert("load", LOAD);
-    intrinsics.insert("load_source_imports", LOAD_SOURCE_IMPORTS);
-    intrinsics.insert("call", CALL);
+    let intrinsics = intrinsics::standard();
 
     // Check for duplicate function arguments.
     let mut arg_names: HashSet<Arc<String>> = HashSet::new();
@@ -804,161 +775,3 @@ impl Kind {
         })
     }
 }
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub enum ArgConstraint {
-    Arg(usize),
-    Return,
-    Default,
-}
-
-#[derive(Debug, Copy, Clone)]
-pub struct Intrinsic {
-    pub arg_constraints: &'static [ArgConstraint],
-    pub returns: bool,
-}
-
-static PRINTLN: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: false
-};
-
-static PRINT: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: false
-};
-
-static CLONE: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: false
-};
-
-static DEBUG: Intrinsic = Intrinsic {
-    arg_constraints: &[],
-    returns: false
-};
-
-static BACKTRACE: Intrinsic = Intrinsic {
-    arg_constraints: &[],
-    returns: false
-};
-
-static SLEEP: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: false
-};
-
-static ROUND: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static RANDOM: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static READ_NUMBER: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static READ_LINE: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static TRIM_RIGHT: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static LEN: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static PUSH: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default, ArgConstraint::Arg(0)],
-    returns: false
-};
-
-static SQRT: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static ASIN: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static SIN: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static COS: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static ACOS: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static TAN: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static ATAN: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static EXP: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static LN: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static LOG2: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static LOG10: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static TO_STRING: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static TYPEOF: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static LOAD: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default],
-    returns: true
-};
-
-static LOAD_SOURCE_IMPORTS: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default; 2],
-    returns: true
-};
-
-static CALL: Intrinsic = Intrinsic {
-    arg_constraints: &[ArgConstraint::Default; 3],
-    returns: true
-};
